front: introduce useCategoryColor
Category colors might be  usefull in several parts of the code making it relevant to extract this logic in a custom hook.
